feat(alerts):Combine banner and inline tests into one
