+++ src/main/java/compounds/fixtures/alerts.html
@@ -35,6 +35,6 @@
 
 <script type="text/javascript" src="../jsfiles/dist.compounds.js"></script>
 <script type="text/javascript" src="../jsfiles/eventInstantiator.compounds.js"></script>
-<script type="text/javascript" src="../jsfiles/bannerAlerts/bannerAlerts.js"></script>
+<script type="text/javascript" src="../jsfiles/alerts/alerts.js"></script>
 </body>
 </html>
