+++ src/main/java/compounds/jsfiles/alerts/alerts.js
